Changed casting from Int to Double for country budget when country with 'max country budget' was searched, although I haven't found any use of the resultant country.

+++ src/platform/modules/search/app/controllers/Application.scala
@@ -15,7 +15,7 @@ object Application extends Controller {
         if(countries.isEmpty) {
           Ok(views.html.search(query, projects.size , projects, countries))
         } else {
-          val country = countries.maxBy(_("countryBudget").asInstanceOf[Int])
+          val country = countries.maxBy(_("countryBudget").asInstanceOf[Double])
           Ok(views.html.search(query, projects.size , projects, countries))
         }
       }
